add log helper

+++ pilipala-frame/pipeline/comment/render.fs
@@ -12,7 +12,7 @@ type ICommentRenderPipelineBuilder =
     abstract Binding: BuilderItem<u64, u64 * CommentBinding>
     abstract CreateTime: BuilderItem<u64, u64 * DateTime>
     abstract UserId: BuilderItem<u64, u64 * u64>
-    abstract Permission: BuilderItem<u64, u64 * u16>
+    abstract Permission: BuilderItem<u64, u64 * u8>
     abstract Item: string -> BuilderItem<u64, u64 * obj>
 
     //用于遍历Item
@@ -23,5 +23,5 @@ type ICommentRenderPipeline =
     abstract Binding: u64 -> u64 * CommentBinding
     abstract CreateTime: u64 -> u64 * DateTime
     abstract UserId: u64 -> u64 * u64
-    abstract Permission: u64 -> u64 * u16
+    abstract Permission: u64 -> u64 * u8
     abstract Item: string -> Option'<u64 -> u64 * obj>
